+++ src/components/tagAnalyzer/board/useTagAnalyzerBoardPanels.ts
@@ -6,10 +6,9 @@ import type {
     TimeRangeMs,
 } from '../domain/time/TimeTypes';
 import {
-    createTimeRangeMs,
-    getTimeRangeCenter,
-    getTimeRangeWidth,
-} from '../domain/time/TimeRangeUtils';
+    getNavigatorTrackPixelWidth,
+    normalizeNavigatorRangeForPanelRange,
+} from './PanelNavigatorRangeLimits';
 import {
     createInitialBoardPanelRecord,
     PanelChartLoadStatus,
@@ -18,50 +17,6 @@ import {
 } from './BoardPanelState';
 import { useBoardPanelChartDataFetching } from './useBoardPanelChartDataFetching';
 import { useBoardPanelRangeMutation } from './useBoardPanelRangeMutation';
-
-const NAVIGATOR_TRACK_SIDE_OFFSET_PX = 56;
-const MIN_NAVIGATOR_SELECTION_PIXEL_WIDTH = 36;
-
-function normalizeNavigatorRangeForPanelRange(
-    panelRange: TimeRangeMs,
-    navigatorRange: TimeRangeMs,
-    navigatorPixelWidth: number,
-): TimeRangeMs {
-    const sNavigatorRange = createTimeRangeMs(
-        Math.min(panelRange.startTime, navigatorRange.startTime),
-        Math.max(panelRange.endTime, navigatorRange.endTime),
-    );
-    const sMaxNavigatorWidth = getMaxNavigatorRangeWidthForMinimumSelection(
-        panelRange,
-        navigatorPixelWidth,
-    );
-
-    if (getTimeRangeWidth(sNavigatorRange) <= sMaxNavigatorWidth) {
-        return sNavigatorRange;
-    }
-
-    const sPanelCenterTime = getTimeRangeCenter(panelRange);
-
-    return createTimeRangeMs(
-        sPanelCenterTime - sMaxNavigatorWidth / 2,
-        sPanelCenterTime + sMaxNavigatorWidth / 2,
-    );
-}
-
-function getMaxNavigatorRangeWidthForMinimumSelection(
-    panelRange: TimeRangeMs,
-    navigatorPixelWidth: number,
-): number {
-    const sPanelWidth = getTimeRangeWidth(panelRange);
-    const sMinimumSelectionRatio =
-        MIN_NAVIGATOR_SELECTION_PIXEL_WIDTH / Math.max(navigatorPixelWidth, 1);
-
-    if (sPanelWidth <= 0 || sMinimumSelectionRatio <= 0) {
-        throw new Error('Cannot normalize navigator range for an invalid panel width.');
-    }
-
-    return Math.max(sPanelWidth, sPanelWidth / Math.min(sMinimumSelectionRatio, 1));
-}
 
 export function useTagAnalyzerBoardPanels({
     panels,
@@ -173,7 +128,7 @@ export function useTagAnalyzerBoardPanels({
         const sChartAreaWidth = getBoardPanelRecord(panelKey).chartAreaWidth;
         const sNavigatorTrackPixelWidth =
             typeof sChartAreaWidth === 'number' && sChartAreaWidth > 0
-                ? Math.max(sChartAreaWidth - NAVIGATOR_TRACK_SIDE_OFFSET_PX, 1)
+                ? getNavigatorTrackPixelWidth(sChartAreaWidth)
                 : undefined;
 
         return sNavigatorTrackPixelWidth === undefined
