+++ src/components/tagAnalyzer/domain/BoardDomain.ts
@@ -59,14 +59,6 @@ export type OverlapPanelInfo = {
 
 export type OverlapShiftDirection = '+' | '-';
 
-export type OverlapOffsetParts = {
-    days: number;
-    hours: number;
-    minutes: number;
-    seconds: number;
-    milliseconds: number;
-};
-
 export type OverlapSelectionChangePayload =
     | {
           panelKey: string;
