+++ src/components/tagAnalyzer/modal/seriesSelection/TagSelectionPanel.tsx
@@ -20,90 +20,8 @@ import {
     mapSelectedSeriesDraftListItems,
     mapTagSearchItemsToListItems,
 } from './tagSelectionPanelHelpers';
-import {
-    DEFAULT_LABEL_STYLE,
-    DEFAULT_TRIGGER_STYLE,
-    MODE_TRIGGER_WRAPPER_STYLE,
-    SELECTED_SERIES_ITEM_STYLE,
-    SELECTED_SERIES_LIST_STYLE,
-} from './TagSelectionConstants';
 import type { TagSelectionPanelViewModel } from './TagSelectionTypes';
-
-const FIELD_ROW_STYLE = {
-    display: 'flex',
-    alignItems: 'center',
-    gap: '12px',
-    width: '100%',
-} as const;
-const PANEL_STACK_STYLE = {
-    display: 'flex',
-    flexDirection: 'column',
-    gap: '12px',
-    width: '100%',
-} as const;
-const ITEM_LIST_ROW_STYLE = {
-    ...FIELD_ROW_STYLE,
-    alignItems: 'flex-start',
-} as const;
-const ITEM_LIST_GROUP_STYLE = {
-    display: 'flex',
-    gap: '12px',
-    flex: '1 1 auto',
-    minWidth: 0,
-} as const;
-const FIELD_LABEL_STYLE = {
-    width: '120px',
-    flexShrink: 0,
-    color: '#c4c4c4',
-    fontSize: '13px',
-    fontWeight: 500,
-} as const;
-const FIELD_CONTROL_GROUP_STYLE = {
-    display: 'flex',
-    alignItems: 'center',
-    gap: '8px',
-    flex: 1,
-    minWidth: 0,
-} as const;
-const FIELD_CONTROL_CONTAINER_STYLE = { flex: 1, minWidth: 0 } as const;
-const JSON_KEY_LABEL_STYLE = {
-    color: '#c4c4c4',
-    fontSize: '13px',
-    fontWeight: 500,
-    flexShrink: 0,
-    display: 'flex',
-    alignItems: 'center',
-    gap: '6px',
-} as const;
-const JSON_KEY_META_STYLE = {
-    color: '#8190a1',
-    fontSize: '10px',
-    fontWeight: 600,
-    lineHeight: 1,
-    whiteSpace: 'nowrap',
-} as const;
-const FIELD_INPUT_STYLE = { height: '30px' } as const;
-const SELECTED_SERIES_TEXT_STYLE = {
-    display: 'flex',
-    flexDirection: 'column',
-    gap: '2px',
-    flex: 1,
-    minWidth: 0,
-} as const;
-const SELECTED_SERIES_SOURCE_STYLE = {
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-    whiteSpace: 'nowrap',
-    color: '#8190a1',
-    fontSize: '11px',
-    fontWeight: 500,
-} as const;
-const SELECTED_SERIES_WARNING_STYLE = {
-    marginTop: '8px',
-    color: '#ff8a3d',
-    fontSize: '12px',
-    lineHeight: '16px',
-} as const;
+import styles from './TagSelectionPanel.module.scss';
 
 function TagSelectionComboboxField({
     label,
@@ -121,10 +39,10 @@ function TagSelectionComboboxField({
     children?: ReactNode;
 }) {
     return (
-        <div style={FIELD_ROW_STYLE}>
-            <label style={FIELD_LABEL_STYLE}>{label}</label>
-            <div style={FIELD_CONTROL_GROUP_STYLE}>
-                <div style={FIELD_CONTROL_CONTAINER_STYLE}>
+        <div className={styles.fieldRow}>
+            <label className={styles.fieldLabel}>{label}</label>
+            <div className={styles.fieldControlGroup}>
+                <div className={styles.fieldControlContainer}>
                     <Combobox.Root
                         options={options}
                         value={value}
@@ -222,7 +140,7 @@ const TagSelectionPanel = ({
     };
 
     return (
-        <div style={PANEL_STACK_STYLE}>
+        <div className={styles.panelStack}>
             <TagSelectionComboboxField
                 label="Table"
                 options={tableOptions}
@@ -247,15 +165,15 @@ const TagSelectionPanel = ({
             >
                 {isJsonValue ? (
                     <>
-                        <span style={JSON_KEY_LABEL_STYLE}>
+                        <span className={styles.jsonKeyLabel}>
                             <span>-&gt;$</span>
                             {selectedJsonKeySummaryLabel ? (
-                                <span style={JSON_KEY_META_STYLE}>
+                                <span className={styles.jsonKeyMeta}>
                                     {selectedJsonKeySummaryLabel}
                                 </span>
                             ) : null}
                         </span>
-                        <div style={FIELD_CONTROL_CONTAINER_STYLE}>
+                        <div className={styles.fieldControlContainer}>
                             <InputSelect
                                 aria-label="JSON key"
                                 type="text"
@@ -269,7 +187,7 @@ const TagSelectionPanel = ({
                                 onSelectChange={onJsonKeySelect}
                                 fullWidth
                                 size="sm"
-                                style={FIELD_INPUT_STYLE}
+                                style={{ height: '30px' }}
                             />
                         </div>
                     </>
@@ -298,10 +216,10 @@ const TagSelectionPanel = ({
                 }
             />
 
-            <div style={ITEM_LIST_ROW_STYLE}>
-                <label style={FIELD_LABEL_STYLE}>Item list</label>
-                <div style={ITEM_LIST_GROUP_STYLE}>
-                    <div style={{ flex: '2 1 0', minWidth: 0 }}>
+            <div className={`${styles.fieldRow} ${styles.itemListRow}`}>
+                <label className={styles.fieldLabel}>Item list</label>
+                <div className={styles.itemListGroup}>
+                    <div className={styles.listColumn}>
                         <List
                             maxHeight={200}
                             items={sAvailableTagListItems}
@@ -322,8 +240,8 @@ const TagSelectionPanel = ({
                         />
                     </div>
 
-                    <div style={{ flex: '2 1 0', minWidth: 0 }}>
-                        <div className={listStyles.list} style={SELECTED_SERIES_LIST_STYLE}>
+                    <div className={styles.listColumn}>
+                        <div className={`${listStyles.list} ${styles.selectedSeriesList}`}>
                             {sSelectedSeriesDraftListItems.length > 0 ? (
                                 <div className={`${listStyles['list__items']} scrollbar-dark`}>
                                     {sSelectedSeriesDraftListItems.map((item) => (
@@ -332,8 +250,7 @@ const TagSelectionPanel = ({
                                             role="button"
                                             tabIndex={0}
                                             title={item.tooltip}
-                                            className={listStyles['list__item']}
-                                            style={SELECTED_SERIES_ITEM_STYLE}
+                                            className={`${listStyles['list__item']} ${styles.selectedSeriesItem}`}
                                             onClick={() => onSelectedSeriesDraftRemove(item.id)}
                                             onKeyDown={(event) =>
                                                 handleSelectedSeriesDraftKeyDown(
@@ -343,20 +260,20 @@ const TagSelectionPanel = ({
                                             }
                                         >
                                             <div className={listStyles['list__item-label']}>
-                                                <div style={{ display: 'flex', alignItems: 'center', gap: '8px', width: '100%' }}>
+                                                <div className={styles.selectedSeriesRow}>
                                                     <span
-                                                        style={SELECTED_SERIES_TEXT_STYLE}
+                                                        className={styles.selectedSeriesText}
                                                         title={item.tooltip}
                                                     >
-                                                        <span style={DEFAULT_LABEL_STYLE}>
+                                                        <span className={styles.selectedSeriesName}>
                                                             {item.selectedSeriesDraft.sourceTagName}
                                                         </span>
-                                                        <span style={SELECTED_SERIES_SOURCE_STYLE}>
+                                                        <span className={styles.selectedSeriesSource}>
                                                             {item.sourceSummary}
                                                         </span>
                                                     </span>
                                                     <div
-                                                        style={MODE_TRIGGER_WRAPPER_STYLE}
+                                                        className={styles.modeTriggerWrapper}
                                                         onClick={(event) => event.stopPropagation()}
                                                     >
                                                         <Dropdown.Root
@@ -372,7 +289,7 @@ const TagSelectionPanel = ({
                                                             <Dropdown.Trigger
                                                                 className="dropdown-trigger-sm"
                                                                 style={{
-                                                                    ...DEFAULT_TRIGGER_STYLE,
+                                                                    width: '100%',
                                                                     ...modeTriggerStyle,
                                                                 }}
                                                             />
@@ -391,10 +308,8 @@ const TagSelectionPanel = ({
                             )}
                         </div>
                         <div
+                            className={styles.selectedCount}
                             style={{
-                                marginTop: '8px',
-                                textAlign: 'right',
-                                fontSize: '12px',
                                 color: sSelectedCountColor,
                             }}
                         >
@@ -404,7 +319,7 @@ const TagSelectionPanel = ({
                             )}
                         </div>
                         {axisKindWarning ? (
-                            <div style={SELECTED_SERIES_WARNING_STYLE}>
+                            <div className={styles.selectedSeriesWarning}>
                                 {axisKindWarning}
                             </div>
                         ) : null}
