+++ src/components/tagAnalyzer/panel/chartBuilder/OptionBuildHelpers/PanelChartSectionOptionBuilder.ts
@@ -74,6 +74,7 @@ export function buildPanelChartLegendOption(
 export function buildPanelChartDataZoomOption(
     display: PanelDisplay,
     panelRange: TimeRangeMs,
+    navigatorSelectionMinValueSpan: number | undefined,
 ): DataZoomComponentOption[] {
     const sPanelRangeDataZoom =
         panelRange.startTime < panelRange.endTime
@@ -95,6 +96,7 @@ export function buildPanelChartDataZoomOption(
             zoomOnMouseWheel: false,
             preventDefaultMouseMove: true,
             disabled: !display.use_zoom,
+            minValueSpan: navigatorSelectionMinValueSpan,
         },
         {
             id: PANEL_SLIDER_DATA_ZOOM_ID,
@@ -109,6 +111,7 @@ export function buildPanelChartDataZoomOption(
             height: PANEL_SLIDER_HEIGHT,
             showDetail: false,
             brushSelect: false,
+            minValueSpan: navigatorSelectionMinValueSpan,
             backgroundColor: 'rgba(0, 0, 0, 0)',
             borderColor: '#7a828c',
             fillerColor: 'rgba(104, 119, 138, 0.28)',
