+++ src/components/tagAnalyzer/panel/chartBuilder/PanelChartRuntimeTypes.ts
@@ -1,5 +1,38 @@
 import type { buildChartSeriesOption } from './OptionBuildHelpers/ChartOptionBuilder';
-import type { EChartDataZoomOptionStateItem } from './ChartInteractionTypes';
+
+export type EChartDataZoomEventItem = {
+    start: number;
+    end: number;
+    startValue?: number;
+    endValue?: number;
+};
+
+export type EChartDataZoomEventPayload =
+    | EChartDataZoomEventItem
+    | {
+          batch: EChartDataZoomEventItem[];
+      };
+
+export type EChartDataZoomOptionStateItem = {
+    start?: number;
+    end?: number;
+    startValue?: number | string | Date;
+    endValue?: number | string | Date;
+};
+
+export type EChartBrushAreaPayload = {
+    coordRange: [number, number] | undefined;
+    range: [number, number] | undefined;
+};
+
+export type EChartBrushPayload = {
+    areas: EChartBrushAreaPayload[] | undefined;
+    batch:
+        | Array<{
+              areas: EChartBrushAreaPayload[] | undefined;
+          }>
+        | undefined;
+};
 
 export type PanelChartBrushOption = {
     brushType: 'lineX' | false;
