+++ src/components/tagAnalyzer/panel/editor/EditorConstants.ts
@@ -2,19 +2,15 @@ import type { CSSProperties } from 'react';
 import InnerLine from '@/assets/image/img_chart_01.png';
 import Scatter from '@/assets/image/img_chart_02.png';
 import Line from '@/assets/image/img_chart_03.png';
-import type { ChartTypeOption, EditTabPanelType } from './EditorTypes';
+import type { EditTabPanelType, EditorChartType } from './EditorTypes';
+
+type ChartTypeOption = {
+    type: EditorChartType;
+    src?: string;
+    alt: string;
+};
 
 export const EDITOR_TABS: EditTabPanelType[] = ['General', 'Data', 'Axes', 'Display', 'Time'];
-
-export const AXES_SECTION_STYLE: CSSProperties = {
-    margin: 0,
-    padding: 0,
-    display: 'flex',
-    flexDirection: 'column',
-    gap: '8px',
-    alignItems: 'start',
-    justifyContent: 'start',
-};
 
 export const EDITOR_X_AXIS_INPUT_STYLE: CSSProperties = {
     width: '150px',
@@ -37,4 +33,5 @@ export const CHART_TYPE_OPTIONS: ChartTypeOption[] = [
     { type: 'Zone', src: InnerLine, alt: 'Zone Chart' },
     { type: 'Dot', src: Scatter, alt: 'Dot Chart' },
     { type: 'Line', src: Line, alt: 'Line Chart' },
+    { type: 'Custom', alt: 'Custom Chart' },
 ];
